+++ src/Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -27,7 +27,7 @@ namespace Roslynator.CSharp
         public const string OrderModifiers = "RCS1019";
         public const string SimplifyNullableOfT = "RCS1020";
         public const string ConvertLambdaExpressionBodyToExpressionBody = "RCS1021";
-        public const string RemoveUnnecessaryBraces = "RCS1031";
+        public const string RemoveUnnecessaryBracesInSwitchSection = "RCS1031";
         public const string RemoveRedundantParentheses = "RCS1032";
         public const string RemoveRedundantBooleanLiteral = "RCS1033";
         public const string RemoveRedundantSealedModifier = "RCS1034";
@@ -207,5 +207,6 @@ namespace Roslynator.CSharp
         public const string NormalizeNullCheck = "RCS1248";
         public const string UnnecessaryNullForgivingOperator = "RCS1249";
         public const string UseImplicitOrExplicitObjectCreation = "RCS1250";
+        public const string RemoveUnnecessaryBraces = "RCS1251";
     }
 }