+++ src/VisualStudioCode/package/src/configurationFiles.generated.ts
@@ -349,7 +349,7 @@ roslynator_analyzers.enabled_by_default = true|false
 # Convert lambda expression body to expression body
 #dotnet_diagnostic.rcs1021.severity = suggestion
 
-# Remove unnecessary braces
+# Remove unnecessary braces in switch section
 #dotnet_diagnostic.rcs1031.severity = silent
 
 # Remove redundant parentheses
@@ -898,6 +898,9 @@ roslynator_analyzers.enabled_by_default = true|false
 # Use implicit/explicit object creation
 #dotnet_diagnostic.rcs1250.severity = none
 # Options: roslynator_object_creation_type_style, roslynator_use_var_instead_of_implicit_object_creation
+
+# Remove unnecessary braces from record declaration
+#dotnet_diagnostic.rcs1251.severity = suggestion
 
 # Use pattern matching
 #dotnet_diagnostic.rcs9001.severity = silent